Simple Test

Test if the input package name etc. was correct / successful. This function is meant to be called from inside the target package.

Usage

uniset_test(uniset_handover)

Arguments

uniset_handover

List length two, containing two elements:

pkgname: The name of the target package.

funcname: The name of the function handing over the required values.

Value

Is printing the parameters defined by the target package, and is returning those parameters in an (invisible) list.

Important

This function is intended to be called from within the target package.

Note

Please refer to uniset for a link to examples and a real-world demo.

Examples

{
if (FALSE) {
# to be called from within the target package
uniset_test(uniset_handover_pkgname, uniset_handover_funcname)
}
}